pyproject.toml
[project]
name = "finalProjectCrm"
version = "0.1.0"
dependencies = [
    "django>=6.0.2",
    "djangorestframework>=3.16.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.12.0",
    "ruff>=0.15.2",
]
